Implemented the /api/v1/chat endpoint in chat.py, added a shared dependencies module to avoid circular imports, and updated the existing API tests to hit the new payload shape and diagnostics error propagation.
