+++ docs/tasks/5-02.md
@@ -1,0 +1,10 @@
+**Implement Chat endpoint**
+  - File: `apps/assistant_api/src/api/v1/chat.py`
+  - **Cursor prompt:**  
+    - Implement `POST /api/v1/chat` accepting `{question, context}` and returning `{answer, citations, diagnostics}`. Handle partial retrieval failures gracefully.
+    - Always write unit tests for the code in a task.
+    - Automatically run unit tests after task is complete.
+    - Automatically run format and style tools by using `.cursor/rules/80-code-style-python.mdc` 
+  - **Definition of Done:** 
+    - endpoint triggers the LangGraph and returns JSON.
+    - all unit tests and linters (ruff, mypy) must pass successfully.
